rename binding to LuxtronicHeatpump
